Bug 560165: Flesh out test projects
Add actual service interfaces and impl classes that match the identifiers
in the uses/provides directives in the module-info files. Also add a main
method in the application project that instantiates a service and calls a
method on the instantiated service. These changes make the projects more
realistic, and allow them to be interactively imported and run.

Signed-off-by: Mike Hogye <stacktracer@gmail.com>