m2e-core issue #271 : Fix integration-tests failing on Windows
The test failures are connected to different file-separators and
line-endings for windows and unix or the required administrator
privileges to create symbolic-links.

In case of missing administrator privileges to create symbolic links,
the error message was made more expressive to guide developers that they
have to run the tests (respectively their IDE) with administrator
privileges.