[TestRepo] add all modules to aggregate and make it work with JDK-11
Add missing modules to the root pom and the downloadsources-aggregator
pom. If you run the Maven deployment to the testrepo from the root of
testrepo-src the entire repo is re-build and not only some modules.

Update the compiler source/target to 1.6 because those seem to be the
minimal possible levels with a JDK-11. And fix the maven-deploy-plugin
version because Maven 3.8.1 uses the deploy-plugin in version 3.0.0-M1
which has breaking changes in its altDeploymentRepository property (and
some projects seem to use always the old version regardless what is
specified in the parent-pom's pluginManagement. Moving it to
deploy-plugin >=3.0.0 would require more changes).

The specification of the compiler's target and source level have to be
changed in 549312-fallbackToLatestEE/pom.xml because the
maven.compiler.source property seem not to override the value from the
parents pluginManagement section.

With those changes the build of the test repo now completes
(nevertheless there are some errors printed, mainly because of missing
checksums)