Increase timeout to wait for a job
The test often "randomly" fails waiting for the job, but using 100ms
seems either a typo or way to short.

This now increases the timeout to 10 seconds, this will make the test
slow in case of real failures but these should not happen anyways in the
usual case. If the test still fails we need to further investigate whats
going on.